+++ python/mosaic/core/library_handler.py
@@ -81,11 +81,26 @@ class MosaicLibraryHandler:
         converters = self.sc._jvm.scala.collection.JavaConverters
 
         JarURI = JavaURI.create("file:" + self._jar_path)
-        lib = JavaJarId(
-            JarURI,
-            ManagedLibraryId.defaultOrganization(),
-            NoVersionModule.simpleString(),
-        )
+        dbr_version = self.spark.conf.get("spark.databricks.clusterUsageTags.sparkVersion").split("-")[0]
+
+        try:
+            # This will fix the exception when running on Databricks Runtime 13.x+
+            optionClass = getattr(self.sc._jvm.scala, "Option$")
+            optionModule = getattr(optionClass, "MODULE$")
+            lib = JavaJarId(
+                JarURI,
+                ManagedLibraryId.defaultOrganization(),
+                NoVersionModule.simpleString(),
+                optionModule.apply(None),
+                optionModule.apply(None),
+            )
+        except:
+            lib = JavaJarId(
+                JarURI,
+                ManagedLibraryId.defaultOrganization(),
+                NoVersionModule.simpleString()
+            )
+
         libSeq = converters.asScalaBufferConverter((lib,)).asScala().toSeq()
 
         context = DatabricksILoop.getSharedDriverContextIfExists().get()