+++ src/main/scala/com/databricks/labs/mosaic/functions/MosaicContext.scala
@@ -965,7 +965,7 @@ object MosaicContext extends Logging {
 
     val tmpDir: String = Files.createTempDirectory("mosaic").toAbsolutePath.toString
 
-    val mosaicVersion: String = "0.3.13"
+    val mosaicVersion: String = "0.3.14"
 
     private var instance: Option[MosaicContext] = None
 
@@ -990,11 +990,11 @@ object MosaicContext extends Logging {
     def reset(): Unit = instance = None
 
     // noinspection ScalaStyle,ScalaWeakerAccess
-    def checkDBR(spark: SparkSession): Unit = {
+    def checkDBR(spark: SparkSession): Boolean = {
         val sparkVersion = spark.conf.get("spark.databricks.clusterUsageTags.sparkVersion", "0")
         val isML = sparkVersion.contains("-ml-")
         val isPhoton = spark.conf.getOption("spark.databricks.photon.enabled").getOrElse("false").toBoolean
-        val isTest = spark.conf.getOption("spark.databricks.clusterUsageTags.clusterType").isEmpty
+        val isTest = !spark.conf.getAll.exists(_._1.startsWith("spark.databricks.clusterUsageTags."))
 
         val dbrMajor = sparkVersion.split("-").head.split("\\.").head.toInt
         if (
@@ -1018,7 +1018,9 @@ object MosaicContext extends Logging {
                          |  Mosaic will stop working on this cluster after v0.3.x.""".stripMargin
             logWarning(msg)
             println(msg)
-
+            false
+        } else {
+            true
         }
     }