+++ src/main/scala/com/databricks/labs/mosaic/gdal/MosaicGDAL.scala
@@ -51,9 +51,7 @@ object MosaicGDAL extends Logging {
         gdal.SetConfigOption("GDAL_PAM_PROXY_DIR", GDAL_PAM_PROXY_DIR)
         gdal.SetConfigOption("GDAL_PAM_ENABLED", "NO")
         gdal.SetConfigOption("CPL_VSIL_USE_TEMP_FILE_FOR_RANDOM_WRITE", "NO")
-        gdal.SetConfigOption("GDAL_CACHEMAX", "64")
         gdal.SetConfigOption("CPL_LOG", s"$CPL_TMPDIR/gdal.log")
-        gdal.SetConfigOption("CPL_DEBUG", "ON")
         mosaicConfig.getGDALConf.foreach { case (k, v) => gdal.SetConfigOption(k.split("\\.").last, v) }
     }